STYLE: Prefer bool rather than implicit type conversion to/from bool
/opt/llvm_trunk/bin/clang-tidy -fix -checks=-*,*bool* -p ~/src/BT-bld/BRAINSTools-build/ ~/src/BRAINSTools/BRAINSABC/brainseg/*.cxx

+++ BRAINSABC/brainseg/ESLR.cxx
@@ -43,7 +43,7 @@ int main(int argc, char *argv[])
   ByteImageType::Pointer myCleanRegion
     = ExtractSingleLargestRegion(low, high, openingSize, closingSize, safetySize, myDirtyRegion);
 
-  if( preserveOutside == true )  // For values outside the specified range,
+  if( static_cast<int>(preserveOutside) == true )  // For values outside the specified range,
                                  // preserve those values.
     {
     std::cout << "PRESERVING OUTSIDE VALUES" << std::endl;
